+++ flutter_inappwebview_ios/lib/src/in_app_webview/in_app_webview_controller.dart
@@ -2233,9 +2233,13 @@ class IOSInAppWebViewController extends PlatformInAppWebViewController
   }
 
   @override
-  Future<void> requestFocus() async {
+  Future<bool?> requestFocus(
+      {FocusDirection? direction,
+        InAppWebViewRect? previouslyFocusedRect}) async {
     Map<String, dynamic> args = <String, dynamic>{};
-    return await channel?.invokeMethod('requestFocus', args);
+    args.putIfAbsent("direction", () => direction?.toNativeValue());
+    args.putIfAbsent("previouslyFocusedRect", () => previouslyFocusedRect?.toMap());
+    return await channel?.invokeMethod<bool>('requestFocus', args);
   }
 
   @override